Rename Mouse and Time events
