+++ Assets/_Scripts/Player/Modules/SceneGraph/MouseEventListenerProxy.cs
@@ -3,17 +3,14 @@ using System.Collections.Generic;
 using UnityEngine;
 using Alice.Tweedle.Interop;
 using Alice.Player.Modules;
-using Alice.Player.Primitives;
 
 namespace Alice.Player.Unity {
-    public class TimeEvent{
+    public class MouseEventListenerProxy{
         
-        public PAction<Duration> listener;
-        public float frequency;
+        public PAction listener;
         public OverlappingEventPolicy policy;
 
         private bool callActive = false;
-        private float lastFireTime = 0f;
         private float lastCheckedTime = 0f;
         private int queuedCalls = 0;
 
@@ -21,14 +18,10 @@ namespace Alice.Player.Unity {
         // TODO: Figure out real time vs simulated time?
         ///////////////////////
 
-        public TimeEvent(PAction<Duration> listener, float freq, OverlappingEventPolicy policy){
+        public MouseEventListenerProxy(PAction listener, OverlappingEventPolicy policy){
             this.listener = listener;
-            this.frequency = freq;
             this.policy = policy;
-        }
-
-        public float getTimeSinceLastFire(){
-            return Time.time - lastFireTime;
+            Debug.Log("Added mouse event");
         }
 
         public void CallEvent(){
@@ -42,9 +35,8 @@ namespace Alice.Player.Unity {
                     return;
                 }
             }
-            lastFireTime = Time.time;
             AsyncReturn callReturn;
-            callReturn = listener.Call(new Duration(lastFireTime));
+            callReturn = listener.Call();
             callActive = true;
             callReturn.OnReturn(() => {
                 returnedCall();
@@ -57,15 +49,6 @@ namespace Alice.Player.Unity {
             if(queuedCalls > 0)
             {
                 queuedCalls--;
-                CallEvent();
-            }
-        }
-
-        public void CheckEvent()
-        {
-            if(Time.time - lastCheckedTime > frequency)
-            {
-                lastCheckedTime = Time.time;
                 CallEvent();
             }
         }
