+++ Assets/_Scripts/Player/Modules/SceneGraph/MouseEventListenerProxy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fb22d6145121ca34eb94443f3cca753a
+guid: dc79e0a9b51bd4f4ca819ca0156eb3a7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
