+++ Assets/_Scripts/Player/Modules/SceneGraph/TimeEventListenerProxy.cs
@@ -3,14 +3,17 @@ using System.Collections.Generic;
 using UnityEngine;
 using Alice.Tweedle.Interop;
 using Alice.Player.Modules;
+using Alice.Player.Primitives;
 
 namespace Alice.Player.Unity {
-    public class MouseEvent{
+    public class TimeEventListenerProxy{
         
-        public PAction listener;
+        public PAction<Duration> listener;
+        public float frequency;
         public OverlappingEventPolicy policy;
 
         private bool callActive = false;
+        private float lastFireTime = 0f;
         private float lastCheckedTime = 0f;
         private int queuedCalls = 0;
 
@@ -18,10 +21,14 @@ namespace Alice.Player.Unity {
         // TODO: Figure out real time vs simulated time?
         ///////////////////////
 
-        public MouseEvent(PAction listener, OverlappingEventPolicy policy){
+        public TimeEventListenerProxy(PAction<Duration> listener, float freq, OverlappingEventPolicy policy){
             this.listener = listener;
+            this.frequency = freq;
             this.policy = policy;
-            Debug.Log("Added mouse event");
+        }
+
+        public float getTimeSinceLastFire(){
+            return Time.time - lastFireTime;
         }
 
         public void CallEvent(){
@@ -35,8 +42,9 @@ namespace Alice.Player.Unity {
                     return;
                 }
             }
+            lastFireTime = Time.time;
             AsyncReturn callReturn;
-            callReturn = listener.Call();
+            callReturn = listener.Call(new Duration(lastFireTime));
             callActive = true;
             callReturn.OnReturn(() => {
                 returnedCall();
@@ -49,6 +57,15 @@ namespace Alice.Player.Unity {
             if(queuedCalls > 0)
             {
                 queuedCalls--;
+                CallEvent();
+            }
+        }
+
+        public void CheckEvent()
+        {
+            if(Time.time - lastCheckedTime > frequency)
+            {
+                lastCheckedTime = Time.time;
                 CallEvent();
             }
         }
